openreview_acceptance_predictor/requirements.txt
openreview-py>=1.46.0
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
sentence-transformers>=2.7.0
transformers>=4.40.0
datasets>=2.19.0
accelerate>=0.30.0
peft>=0.11.0
bitsandbytes>=0.43.0
trl>=0.8.6
torch>=2.2.0
pypdf>=4.2.0
requests>=2.31.0
tqdm>=4.66.0
joblib>=1.4.0
faiss-cpu>=1.8.0
matplotlib>=3.8.0
seaborn>=0.13.0
python-dotenv>=1.0.0
